Chapter 1.b: Add VarHandle subword CAS tests

+++ tests-chapter-0a/src/main/java/org/openjdk/jcstress/tests/defaultValues/arrays/large/volatiles/ByteTest.java
@@ -55,7 +55,7 @@ public class ByteTest {
         } else {
             boolean allDefault = true;
             for (byte v : a) {
-                allDefault &= (v == 0);
+                allDefault &= (v == (byte)0);
             }
             r.r1 = allDefault ? 0 : 1;
         }
